+++ spicedmoon/__init__.py
@@ -1,6 +1,6 @@
 from .geoselenic import (
     get_moon_datas_from_extra_kernels,
-    get_moon_datas_xyzs_no_zenith_azimuth,
+    get_moon_datas_xyzs,
 )
 from .geotic import get_moon_datas
 from .selenic import get_moon_datas_from_moon